stage release

+++ lib/utils/storage.util.ts
@@ -1,12 +1,13 @@
+import { DraftBookingForm } from '../types/storage.type'
 import { Seat } from '@/lib/types/seat.type'
 import { getShowTime } from '@/lib/utils/movie.util'
 import { seatLayout } from '@/configs/movie.config'
 
 export const saveDraftBookingForm = (
   label: string,
-  value: string | Record<number, Seat[]>
+  value: string | number | Record<number, Seat[]> | string[]
 ) => {
-  if (!label || !value || Object.keys(value).length <= 0) return
+  if (!label) return
 
   const draftBookingForm = JSON.parse(
     localStorage.getItem('draftBookingForm') || '{}'
@@ -17,15 +18,31 @@ export const saveDraftBookingForm = (
   localStorage.setItem('draftBookingForm', JSON.stringify(draftBookingForm))
 }
 
-export const saveSeatLayoutsAndShowTimeForMovies = (movieId: number) => {
+export const saveSeatLayoutsAndShowTimeForMovies = (movieId: string) => {
+  const existingMovieDetails = localStorage.getItem(`${movieId}`)
+
+  if (existingMovieDetails) return
+
   localStorage.setItem(
     `${movieId}`,
-    JSON.stringify({ seatLayout, showTime: getShowTime(movieId) })
+    JSON.stringify({ seatLayout, showTime: getShowTime(Number(movieId)) })
   )
 }
 
 export const getDraftBookingForm = () =>
   JSON.parse(localStorage.getItem('draftBookingForm') || '{}')
+
+export const getDraftBookingFormPropertyValue = (
+  property: keyof DraftBookingForm
+) => {
+  const draftBookingForm: DraftBookingForm = JSON.parse(
+    localStorage.getItem('draftBookingForm') || '{}'
+  )
+
+  return draftBookingForm[property] !== undefined
+    ? draftBookingForm[property]
+    : null
+}
 
 export const getSeatLayoutForMovie = (movieId: string) => {
   const movieDetails = JSON.parse(localStorage.getItem(movieId) || '{}')
